feat: optional restart in case of read/write failure
The config allows respawning a server together with
Mongo or RabbitMQ connections in case when fetching from
MongoDB or publishing to RabbitMQ fail.

By default, it is off since most of users would probably
run the container in Docker with always-restart or similar.

The config does not apply to the connetivity issues on
clients creation.

The config is per collection.